Fix bug in `lua-eldoc-mode-help-at-point`
In some situations, such as when `point` is on the first line of the
buffer before any Lua code, the function
`lua-eldoc-mode-thing-at-point-no-properties` may return NIL.

This causes the call to `s-contains?" in
`lua-eldoc-mode-help-at-point` to crash with an error.

Added a new unit test to lua-eldoc-mode-test.el that tests this bugfix.